Adapt unsigned constant extraction to BigInt and the sym namespace. Collect unsigned ranges once during traversal, separately from signed equations, and cover high-bit constants and modular predicates.

Validation: full local build; 1103 Python tests passed, 53 skipped, 8 xfailed; 7 local C++ probes passed; changed-file pre-commit passed. TestTileAwareCompaction is excluded from the final run after reproducing its reflection-constructor failure with upstream conditional_bounds.cc. All four unsigned regressions fail with upstream code and pass with this fix.

+++ src/relax/backend/vm/exec_builder.cc
@@ -351,7 +351,7 @@ TVM_FFI_STATIC_INIT_BLOCK() {
            [](ExecBuilder builder, ffi::String name, ffi::Array<IntImm> args, int64_t dst) {
              std::vector<Instruction::Arg> args_;
              for (size_t i = 0; i < args.size(); ++i) {
-               args_.push_back(Instruction::Arg::FromData(args[i]->value));
+               args_.push_back(Instruction::Arg::FromData(static_cast<int64_t>(args[i]->value)));
              }
              auto dst_ = Instruction::Arg::Register(dst);
              builder->EmitCall(name, args_, dst_.value());
